+++ Dockerfile.dev
@@ -82,6 +82,8 @@ COPY packages/httpsPortal/package.json \
   packages/httpsPortal/
 COPY packages/dyndns/package.json \
   packages/dyndns/
+COPY packages/upnpc/package.json \
+  packages/upnpc/
 RUN yarn --frozen-lockfile --non-interactive --ignore-optional
 
 # Build params
@@ -141,6 +143,12 @@ RUN yarn build
 # Build dockerApi
 WORKDIR /app/packages/dockerApi/
 COPY packages/dockerApi/ .
+RUN yarn build
+# Results in dist/*
+
+# Build upnpc
+WORKDIR /app/packages/upnpc/
+COPY packages/upnpc/ .
 RUN yarn build
 # Results in dist/*
