+++ packages/dappmanager/package.json
@@ -21,6 +21,7 @@
   },
   "license": "GPL-3.0",
   "dependencies": {
+    "@dappnode/upnpc": "^0.1.0",
     "@dappnode/dyndns": "^0.1.0",
     "@dappnode/common": "^0.1.0",
     "@dappnode/db": "^0.1.0",
