+++ packages/dockerApi/src/index.ts
@@ -10,3 +10,4 @@ export * from "./cleanOldImages.js";
 export * from "./list/index.js";
 export * from "./compose/index.js";
 export * from "./api/index.js";
+export { getDappmanagerImage } from "./getDappmanagerImage.js";
