+++ packages/upnpc/.eslintrc.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  extends: [
+    "../../.eslintrc.cjs", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
+  ],
+}
