+++ packages/upnpc/.mocharc.yaml
@@ -1,0 +1,8 @@
+colors: true
+exit: true
+extension: [ts]
+require:
+  - dotenv/config
+node-option:
+  - experimental-specifier-resolution=node
+  - loader=ts-node/esm
