+++ packages/upnpc/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dappnode/dockerapi",
+  "name": "@dappnode/upnpc",
   "type": "module",
   "version": "0.1.0",
   "license": "GPL-3.0",
@@ -14,31 +14,21 @@
   },
   "scripts": {
     "build": "tsc -p tsconfig.json",
-    "test": "TEST=true mocha --config ./.mocharc.yaml --recursive ./test/unit",
+    "test": "mocha --config ./.mocharc.yaml --recursive ./test/unit",
     "dev": "tsc -w",
     "lint": "eslint . --ext .ts --fix src"
   },
   "dependencies": {
     "@dappnode/common": "^0.1.0",
-    "@dappnode/hostscriptsservices": "^0.1.0",
-    "@dappnode/params": "^0.1.0",
-    "@dappnode/types": "^0.1.30",
     "@dappnode/utils": "^0.1.0",
-    "@types/lodash-es": "^4.17.9",
-    "dargs": "7.0.0",
-    "dockerode": "^3.2.1",
-    "lodash-es": "^4.17.21",
-    "memoizee": "0.4.14",
-    "tar-stream": "2.1.4"
+    "@dappnode/dockerapi": "^0.1.0",
+    "is-ip": "3.0.0"
   },
   "devDependencies": {
     "@types/chai": "^4.3.6",
-    "@types/dockerode": "^3.2.1",
-    "@types/memoizee": "0.4.4",
-    "@types/mocha": "^10.0.2",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^20.6.2",
-    "@types/tar-stream": "2.1.0",
-    "chai": "^4.3.10",
+    "chai": "^4.3.8",
     "mocha": "^10.2.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
