+++ packages/upnpc/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    /* Modules */
+    "types": ["node", "mocha"],
+
+    /* Emit */
+    "outDir": "dist",
+    "sourceMap": true,
+
+    /* Language and Environment */
+    "lib": ["ESNext"]
+  },
+  "ts-node": { "esm": true, "experimentalSpecifierResolution": "node" },
+  "include": ["src/**/*", "src/**/*.json"],
+  "exclude": ["node_modules", "test/**/*"]
+}
